ENH: Enhanced the behaviour of download behavior in DeepInfer.
Direct download of large docker images take a while. Slicer application
is single threaded and it locks during the download progress update frozen
and the slicer becomes irresponsive.
Now instead of direct download a popup window shows up with the docker
pull command. User have to copy it and paste it into terminal.